fontconfig_parser::types

Module dir

Macros

define_calculate_path 🔒

Structs

CacheDir

Dir

Include

RemapDir

This element contains a directory name where will be mapped as the path ‘as-path’ in cached information. This is useful if the directory name is an alias (via a bind mount or symlink) to another directory in the system for which cached font information is likely to exist. ‘salt’ property affects to determine cache filename as same as Dir element.

Enums

DirPrefix

PrefixBehavior

Functions

config_get_file_name 🔒

Given a relative path to a config file, this function returns the complete file name to load.

config_home 🔒

Get the location to user home directory.

expand_tilde 🔒